extract normalization model training to the proper place in the processors architecture
