+++ processors/base/normalizers/__init__.py
@@ -4,8 +4,4 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from . import processors
-from . import writers
-from . import helpers
-from . import readers
-from . import normalizers
+from .location_normalizer import get_normalized_form
